std: associative container API for contains/at
