+++ src/hstd/stdlib/ContainerAPI.hpp
@@ -2,6 +2,7 @@
 
 #include <hstd/system/aux_templates.hpp>
 #include <hstd/system/aux_utils.hpp>
+#include <hstd/stdlib/Vec.hpp>
 
 namespace hstd {
 template <typename Derived, typename Container, typename ValueType>
@@ -35,6 +36,23 @@ struct AssociativeContainerAdapterBase : CRTP_this_method<Derived> {
     bool contains(KeyType const& key) const {
         return _this()->contains_impl(key);
     }
+
+    ValueType const& at(KeyType const& key) {
+        return _this()->at_impl(key);
+    }
+
+    KeyType const& get_pair_key(
+        std::pair<KeyType, ValueType> const& pair) const {
+        return pair.first;
+    }
+
+    Vec<KeyType> keys() const {
+        Vec<KeyType> result;
+        for (auto it = _this()->begin(); it != _this()->end(); ++it) {
+            result.push_back(_this()->get_pair_key(*it));
+        }
+        return result;
+    }
 };
 
 template <typename T>
@@ -55,4 +73,34 @@ concept HasAssociativeContainerAdapter = requires {
     typename AssociativeContainerAdapter<T>::pair_key_type;
     typename AssociativeContainerAdapter<T>::pair_value_type;
 };
+
+
+template <typename T>
+class SequentialContainerAdapter<hstd::Vec<T>>
+    : public SequentialContainerAdapterBase<
+          SequentialContainerAdapter<hstd::Vec<T>>,
+          hstd::Vec<T>,
+          T> {
+  public:
+    using container_type  = hstd::Vec<T>;
+    using item_value_type = T;
+
+    container_type const* container;
+
+    SequentialContainerAdapter(const container_type* container) {}
+
+    auto begin_impl() const { return container->begin(); }
+    auto end_impl() const { return container->end(); }
+
+    void add_impl(const item_value_type& value) {
+        const_cast<container_type*>(container)->push_back(value);
+    }
+
+    int size_impl() const { return static_cast<int>(container->size()); }
+
+    void begin_insert_impl() {}
+    void end_insert_impl() {}
+    void clear_impl() { const_cast<container_type*>(container)->clear(); }
+};
+
 } // namespace hstd
